wordbankr: Accessing the Wordbank Database

Connecting to Wordbank, an open repository for developmental vocabulary data. For more information on the underlying data, see http://wordbank.stanford.edu.

See also

Author

Maintainer: Mika Braginsky mika.br@gmail.com

Other contributors:

Daniel Yurovsky yurovsky@stanford.edu [contributor]

Michael Frank mcfrank@stanford.edu [contributor]

Danielle Kellier dkellier@stanford.edu [contributor]

Alvin Tan tanawm@stanford.edu [contributor]